+++ NEWS.md
@@ -9,6 +9,8 @@ Architectural changes:
 
 Bug fixes:
 
+- Preprocessor: automatically detect CGNS file type (ADF or HDF5).
+
 - Preprocessor: fix crash when reading CGNS files with NGON (polyhedra)
   sections and vertex-based boundary condition definitions.